tetragon: Add max flag to KProbeArg selector
At the moment argument types like char_buf and cha_iovec will
retrieve only certain amount of data dues to the event size
limitations.

Adding 'max' boolean flag that instructs the code to get as
much as possible data of the argument.

The actual implementation comes in following changes, this is
just the tracing policy support.

Signed-off-by: Jiri Olsa <jolsa@kernel.org>