tetragon: Refactor createGenericKprobeSensor function
Curve out addKprobe which adds single kprobe, so we can pass
specific name to it following changes and support symbols lists.

The number of in/out arguments for addKprobe function is too
big so I added in and out structs to carry it.

There's no functional change intended, just preparation for
following changes.

Signed-off-by: Jiri Olsa <jolsa@kernel.org>