tetragon: Add InMap operator range test
Adding test for InMap operator with range value for both kprobes
and tracepoints.

The kprobe tests hooks sys_dup syscall and filters for fd in
range 9910 to 9920 while the test generate wider range.

The tracepoint test hooks raw tracepoint and uses InMap operator
to filter for dup syscall and for same fd range as kprobe test.

Signed-off-by: Jiri Olsa <jolsa@kernel.org>