tetragon: Add PinPath to Map object
Adding PinPath to Map object to carry path relative to the sysfs
bpf root tetragon tree.

At the moment we have map's Name as a real (bpf object) name and
PinName when we need to pin map under different name.

The PinName will be removed once we move to new hierarchy structure,
but we still need to keep the relative pinned path of the map.

Also we need to keep the Map pointer in PinMap map, becase we need
to access PinPath in loader code.

Signed-off-by: Jiri Olsa <jolsa@kernel.org>