tetragon/windows:  Support Windows create and exit process - sensor changes
This PR lists sensor side changes to support CreateProcess and ExitProcess events.
The program loader needs to load the entire collection of native Windows ebps program image using cilium/ebpf library.
The specs for a native Windows bpf program is not available as it is not in ELF format
This changes the order in which maps are loaded - collection is loaded first which loads maps and programs automatically.

Signed-off-by: Anadi Anadi<aanadi@cisco.com>